scripts/requirements.txt
bs4
click
coloredlogs
flickrapi
html2text
html5lib
imdbpy
pillow
pycryptodome
pyyaml
requests
tqdm
